+++ changelog.md
@@ -11,6 +11,7 @@
 * Leftover Qt5 fixes
 * Fix fa updater not cancelling download (#802, #803)
 * Fix regression - map and mod previews not downloading correctly (#804, #805)
+* Fix maps with spaces in them sometimes failing to download in vault (#771, #772)
 
 Contributors:
  - Wesmania